Refactor HR module into dedicated pages
